Add more examples to playground
