+++ package.json
@@ -87,6 +87,7 @@
     "eslint-plugin-standard": "^4.0.1",
     "exit-hook": "^2.2.0",
     "html": "1.0.0",
+    "html-loader": "^1.1.0",
     "html-webpack-plugin": "^4.3.0",
     "http-server": "^0.12.3",
     "husky": "^4.2.5",
