+++ playground/page-field-removal/tmpl.html
@@ -10,5 +10,9 @@
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" crossorigin="anonymous">
 </head>
 <body>
+    <%= require('html-loader!../partials/nav.html') %>
+
     <div id="app"></div>
+
+    <%= require('html-loader!../partials/footer.html') %>
 </body>
