+++ playground/partials/footer.html
@@ -1,0 +1,22 @@
+<div class="bg-light py-5">
+  <div class="container ">
+    <ul class="nav">
+      <li class="nav-item">
+        <a class="nav-link"
+                href="https://github.com/ivarprudnikov/rjsf-conditionals">
+          Source code / Github repository
+        </a>
+      </li>
+      <li class="nav-item">
+        <a class="nav-link" href="https://github.com/ivarprudnikov/rjsf-conditionals/blob/master/README.md">
+          Docs
+        </a>
+      </li>
+      <li class="nav-item">
+        <a class="nav-link" href="https://github.com/rjsf-team/react-jsonschema-form">
+          Depends on React JSON Schema Form
+        </a>
+      </li>
+    </ul>
+  </div>
+</div>
