+++ playground/partials/nav.html
@@ -1,0 +1,34 @@
+<nav class="navbar navbar-expand navbar-dark bg-primary fixed-top">
+    <a class="navbar-brand" href="index.html">
+        rjsf-conditionals
+    </a>
+    <ul class="navbar-nav ml-auto">
+        <li class="nav-item">
+            <a class="nav-link" href="index.html">
+                BMI calculator
+            </a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" href="sum.html">
+                Simple sum
+            </a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" href="field-removal.html">
+                Toggle field
+            </a>
+        </li>
+    </ul>
+    <ul class="navbar-nav ml-3">
+        <li class="nav-item">
+            <a class="nav-link" href="https://github.com/ivarprudnikov/rjsf-conditionals">
+                Github repository
+            </a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" href="https://github.com/ivarprudnikov/rjsf-conditionals/blob/master/README.md">
+                Docs
+            </a>
+        </li>
+    </ul>
+</nav>
