+++ webpack.playground.dev.js
@@ -8,8 +8,14 @@ module.exports = {
   // https://webpack.js.org/configuration/devtool/
   devtool: 'eval-cheap-module-source-map',
 
-  entry: './playground/app/index.js',
+  // https://webpack.js.org/concepts/entry-points/#multi-page-application
+  entry: {
+    index: './playground/page-index/index.js',
+    sum: './playground/page-sum/index.js',
+    'field-removal': './playground/page-field-removal/index.js'
+  },
 
+  // https://webpack.js.org/concepts/loaders/
   module: {
     rules: [
       {
@@ -38,7 +44,21 @@ module.exports = {
   },
 
   // https://webpack.js.org/plugins/html-webpack-plugin/
-  plugins: [new HtmlWebpackPlugin({
-    template: './playground/tmpl.html'
-  })]
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: './playground/page-index/tmpl.html',
+      chunks: ['index'],
+      filename: 'index.html'
+    }),
+    new HtmlWebpackPlugin({
+      template: './playground/page-sum/tmpl.html',
+      chunks: ['sum'],
+      filename: 'sum.html'
+    }),
+    new HtmlWebpackPlugin({
+      template: './playground/page-field-removal/tmpl.html',
+      chunks: ['field-removal'],
+      filename: 'field-removal.html'
+    })
+  ]
 };
